refactor : 역할이 모호한 StringAddCalculator 삭제
