+++ src/main/java/lotto/step1/Application.java
@@ -6,7 +6,8 @@ public class Application {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         String text = sc.nextLine();
-        StringAddCalculator stringAddCalculator = new StringAddCalculator(text);
-        System.out.println(stringAddCalculator.splitAndSum());
+        Calculator calculator = new Calculator(text);
+        int result = calculator.calculate();
+        System.out.println(result);
     }
 }
